+++ tests/test_aggregator.py
@@ -6,64 +6,151 @@
 # __maintainer__ = "Ronie Martinez"
 # __email__ = "ronmarti18@gmail.com"
 # __status__ = "Development"
+import string
+
+import TexSoup
+
 from latex2mathml.aggregator import aggregate
 
 
-def test_single_group():
-    assert [['a']] == aggregate('{a}')
+def test_alphabets():
+    alphabets = string.ascii_letters
+    assert list(alphabets) == aggregate(TexSoup.TexSoup(alphabets))
 
 
-def test_multiple_groups():
-    assert [['a'], ['b']] == aggregate('{a}{b}')
+def test_numbers():
+    numbers = string.digits
+    assert [numbers] == aggregate(TexSoup.TexSoup(numbers))
 
 
-def test_inner_group():
-    assert [['a', '+', ['b']]] == aggregate('{a+{b}}')
+def test_numbers_with_decimals():
+    decimal = '12.56'
+    assert [decimal] == aggregate(TexSoup.TexSoup(decimal))
 
 
-def test_subscript():
-    assert ['_', 'a', 'b'] == aggregate('a_b')
+def test_numbers_and_alphabets():
+    s = '5x'
+    assert list(s) == aggregate(TexSoup.TexSoup(s))
 
 
-def test_superscript():
-    assert ['^', 'a', 'b'] == aggregate('a^b')
+def test_decimals_and_alphabets():
+    s = '5.8x'
+    assert ['5.8', 'x'] == aggregate(TexSoup.TexSoup(s))
 
 
-def test_subscript_and_superscript():
-    assert ['_^', 'a', 'b', 'c'] == aggregate('a_b^c')
+def test_string_with_spaces():
+    s = '3 x'
+    assert ['3', 'x'] == aggregate(TexSoup.TexSoup(s))
 
 
-def test_root():
-    assert [r'\root', ['2'], ['3']] == aggregate(r'\sqrt[3]{2}')
+def test_operators():
+    s = '+-*/=()[]_^{}'
+    assert list(s) == aggregate(TexSoup.TexSoup(s))
+
+
+def test_numbers_alphabets_and_operators():
+    s = '3 + 5x - 5y = 7'
+    assert ['3', '+', '5', 'x', '-', '5', 'y', '=', '7'] == aggregate(TexSoup.TexSoup(s))
+
+
+def test_symbols():
+    s = r'\alpha\beta'
+    assert [r'\alpha', r'\beta'] == aggregate(TexSoup.TexSoup(s))
+
+
+def test_symbols_appended_number():
+    s = r'\frac2x'
+    assert [r'\frac', '2', 'x'] == aggregate(TexSoup.TexSoup(s))
 
 
 def test_matrix():
-    assert [r'\matrix', [['a', 'b'], ['c', 'd']]] == list(aggregate(r'\begin{matrix}a & b \\ c & d \end{matrix}'))
+    s = r'\begin{matrix}a & b \\ c & d \end{matrix}'
+    assert [r'\matrix', '{', 'a', '&', 'b', r'\\', 'c', '&', 'd', '}'] == aggregate(TexSoup.TexSoup(s))
+
+
+def test_matrix_with_alignment():
+    s = r'\begin{matrix*}[r]a & b \\ c & d \end{matrix*}'
+    assert [r'\matrix*', 'r', '{', 'a', '&', 'b', r'\\', 'c', '&', 'd', '}'] == aggregate(TexSoup.TexSoup(s))
+
+
+def test_matrix_with_negative_sign():
+    s = r'\begin{matrix}-a & b \\ c & d \end{matrix}'
+    assert [r'\matrix', '{', '-', 'a', '&', 'b', r'\\', 'c', '&', 'd', '}'] == aggregate(TexSoup.TexSoup(s))
+
+
+def test_simple_array():
+    s = r'\begin{array}{cc} 1 & 2 \\ 3 & 4 \end{array}'''
+    assert [r'\array', 'cc', '{', '1', '&', '2', r'\\', '3', '&', '4', '}'] == aggregate(TexSoup.TexSoup(s))
+
+
+def test_single_group():
+    assert [['a']] == aggregate(TexSoup.TexSoup('{a}'))
+
+
+def test_multiple_groups():
+    assert [['a'], ['b']] == aggregate(TexSoup.TexSoup('{a}{b}'))
+
+
+def test_inner_group():
+    assert [['a', '+', ['b']]] == aggregate(TexSoup.TexSoup('{a+{b}}'))
+
+
+def test_subscript():
+    assert ['_', 'a', 'b'] == aggregate(TexSoup.TexSoup('a_b'))
+
+
+def test_superscript():
+    assert ['^', 'a', 'b'] == aggregate(TexSoup.TexSoup('a^b'))
+
+
+def test_subscript_and_superscript():
+    assert ['_^', 'a', 'b', 'c'] == aggregate(TexSoup.TexSoup('a_b^c'))
+
+
+def test_root():
+    assert [r'\root', ['2'], ['3']] == aggregate(TexSoup.TexSoup(r'\sqrt[3]{2}'))
+
+
+def test_matrix():
+    assert [r'\matrix', [['a', 'b'], ['c', 'd']]] == \
+           list(aggregate(TexSoup.TexSoup(r'\begin{matrix}a & b \\ c & d \end{matrix}')))
 
 
 def test_matrix_with_alignment():
     assert [r'\matrix*', 'r', [['a', 'b'], ['c', 'd']]] == \
-           list(aggregate(r'\begin{matrix*}[r]a & b \\ c & d \end{matrix*}'))
+           list(aggregate(TexSoup.TexSoup(r'\begin{matrix*}[r]a & b \\ c & d \end{matrix*}')))
 
 
 def test_matrix_with_negative_sign():
     assert [r'\matrix', [[['-', 'a'], 'b'], ['c', 'd']]] == \
-           list(aggregate(r'\begin{matrix}-a & b \\ c & d \end{matrix}'))
+           list(aggregate(TexSoup.TexSoup(r'\begin{matrix}-a & b \\ c & d \end{matrix}')))
 
 
 def test_complex_matrix():
-    assert ['\\matrix', [['_', 'a', ['1'], '_', 'b', ['2']], ['_', 'c', ['3'], '_', 'd', ['4']]]] == \
-           list(aggregate(r'\begin{matrix}a_{1} & b_{2} \\ c_{3} & d_{4} \end{matrix}'))
+    # assert ['\\matrix', [['_', 'a', ['1']], ['_', 'b', ['2']], ['_', 'c', ['3']], ['_', 'd', ['4']]]]
+    assert [r'\matrix', [['_', 'a', ['1'], '_', 'b', ['2']], ['_', 'c', ['3'], '_', 'd', ['4']]]] == \
+           list(aggregate(TexSoup.TexSoup(r'\begin{matrix}a_{1} & b_{2} \\ c_{3} & d_{4} \end{matrix}')))
 
 
 def test_simple_array():
     assert [r'\array', 'cc', [['1', '2'], ['3', '4']]] == \
-           list(aggregate(r'\begin{array}{cc} 1 & 2 \\ 3 & 4 \end{array}'''))
+           list(aggregate(TexSoup.TexSoup(r'\begin{array}{cc} 1 & 2 \\ 3 & 4 \end{array}')))
 
 
 def test_frac():
-    assert ['\\frac', ['1'], ['2']] == list(aggregate(r'\frac{1}{2}'))
+    assert [r'\frac', ['1'], ['2']] == list(aggregate(TexSoup.TexSoup(r'\frac{1}{2}')))
 
 
 def test_over():
-    assert ['\\frac', ['1'], ['2']] == list(aggregate(r'1 \over 2'))
+    assert [r'\frac', ['1'], ['2']] == list(aggregate(TexSoup.TexSoup(r'1 \over 2')))
+
+
+def test_null_delimiter():
+    assert [r'\left\{', r'\right.'] == list(aggregate(TexSoup.TexSoup(r'\left\{\right.')))
+    latex = r'\left\{ \begin{array} { l } { 3x - 5y + 4z = 0} \\ { x - y + 8z = 0} \\ { 2x - 6y + z = 0} \end{array}' \
+            r' \right.'
+    assert [r'\left\{', r'\array', ['l'],
+            [[['3', 'x', '-', '5', 'y', '+', '4', 'z', '=', '0']],
+             [['x', '-', 'y', '+', '8', 'z', '=', '0']],
+             [['2', 'x', '-', '6', 'y', '+', 'z', '=', '0']]],
+            r'\right.'] == list(aggregate(TexSoup.TexSoup(latex)))